brain_system/requirements.txt

langgraph
langchain
langchain-core
langchain-google-genai
langchain-openai
langchain-community
python-dotenv
flask
PyPDF2
